requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
schedule
gunicorn
